major reorganization; added setuptools and sphinx docs folder

+++ tests/ft_apps/make_sides.py
@@ -6,7 +6,7 @@ import time
 import sys
 
 parser = argparse.ArgumentParser()
-parser.add_argument('--sleep', type=int, default=0)
+parser.add_argument('--sleep', type=float, default=0)
 parser.add_argument('--retcode', type=int, default=0)
 args = parser.parse_args()
